Fix #298 — Extract duplicated buildTestApp helpers into AppFactory
Create spec/support/utils/AppFactory.js with three named builder
functions (buildRequestHandlerApp, buildRouteRegisterApp, buildRouterApp)
extracted verbatim from the three spec files, and update each spec to
import and use its corresponding named function.
